Further cleanups to Day 16 Part 1. Still runs fast, about 0.2 seconds (release build - debug build is 10x slower!)
Now that I've looked at other peoples solutions, I'm realizing the reason mine is so complicated is that I ignored the solution of "just cache everything". This was sooort of intentional, though also sort of because I didn't think of it lol. But whatever, I'll commit to solving this problem without a huge memory footprint and keep my solution cacheless, it's a fun challenge.